requirements.txt
uagents==0.22.3
uagents-adapter==0.2.1
langchain-openai==0.3.14
langchain-community==0.3.21
langgraph==0.3.31
python-dotenv==1.0.0
pymongo==4.6.1
langchain-core>=0.1.20
langchain-mcp-adapters>=0.0.14
mcp>=0.0.4
mongoengine>=0.27.0
openai>=1.6.0
requests>=2.31.0
httpx>=0.24.0
aiohttp>=3.8.5
websockets>=11.0.3 